src/main/java/uk/gov/hmcts/reform/waworkflowapi/external/taskservice/TaskClientService.java
package uk.gov.hmcts.reform.waworkflowapi.external.taskservice;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;

import java.util.List;
import java.util.Map;

@Component
public class TaskClientService {
    private final CamundaClient camundaClient;
    private final AuthTokenGenerator authTokenGenerator;
    public static final String WA_TASK_INITIATION_DECISION_TABLE_NAME = "wa-task-initiation";

    @Autowired
    public TaskClientService(@Autowired CamundaClient camundaClient,
                             AuthTokenGenerator authTokenGenerator) {
        this.camundaClient = camundaClient;
        this.authTokenGenerator = authTokenGenerator;
    }

    public void sendMessage(SendMessageRequest sendMessageRequest) {
        camundaClient.sendMessage(
            authTokenGenerator.generate(),
            sendMessageRequest
        );
    }

    public List<Map<String,DmnValue<?>>> evaluate(EvaluateDmnRequest evaluateDmnRequest, String key) {
        return camundaClient.evaluateDmn(
            authTokenGenerator.generate(),
            key,
            evaluateDmnRequest
        );
    }


}
